Move camel/snake helper fns to utils.helpers
