+++ orchestrator/utils/helpers.py
@@ -10,6 +10,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+import re
 from ipaddress import IPv4Address, IPv4Network, IPv6Address, IPv6Network
 from typing import Any, Callable, TypeVar, Union
 
@@ -93,3 +94,12 @@ def to_snake(s: str) -> str:
 
 def is_ipaddress_type(v: Any) -> bool:
     return isinstance(v, (IPv4Address, IPv6Address, IPv4Network, IPv6Network))
+
+
+def snake_to_camel(s: str) -> str:
+    return "".join(x.title() for x in s.split("_"))
+
+
+def camel_to_snake(s: str) -> str:
+    name = re.sub("(.)([A-Z][a-z]+)", r"\1_\2", s)
+    return re.sub("([a-z0-9])([A-Z])", r"\1_\2", name).lower()
